fake_automate/requirements.txt
pyModbusTcp
python-dotenv